move files under data to priv to comply with elixir project req
