+++ priv/pinyin_tone.txt
@@ -1,0 +1,28 @@
+ā,a,1
+á,a,2
+ǎ,a,3
+à,a,4
+ē,e,1
+é,e,2
+ě,e,3
+è,e,4
+ī,i,1
+í,i,2
+ǐ,i,3
+ì,i,4
+ō,o,1
+ó,o,2
+ǒ,o,3
+ò,o,4
+ǖ,v,1
+ǘ,v,2
+ǚ,v,3
+ǜ,u,4
+ū,u,1
+ú,u,2
+ǔ,u,3
+ù,u,4
+üē,ue,1
+üé,ue,2
+üě,ue,3
+üè,ue,4
